requirements.txt
scipy
numpy
pytorch
torchvision
networkx
scikit-learn
ipython
matplotlib
tensorflow==1.13.1
